'cat' functionality now can construct columns
It is now possible to call thresh without any input files. This
could be useful for constructing columns from python statements.

it is now possible to construct columns from arbitrary python
statements. However, it still raises an exception when you try to
use 'cat' because it's not ready for prime time.